Shorten the error message when the crate can't be parsed.
Usually this happen because of a syntax error in the crate. When this
happens, the real error will be shown in the next step while compiling
the crate. However, this warning is so huge that it fills the console and
hides the actual error. (It contains all the token of the crate.)
Strip the very long error message coming from synom.