Allow local type in the rust! signature
The problem is that it is not possible to use, in the rust! macro, types
declared in the same cpp! macro, because the extern "C" declaration happens
before. This is actually quite frequent as we would call the #include and
then inherit from class in the same cpp! macro.
(As reported in https://github.com/woboq/qmetaobject-rs/issues/4 )

The solution I used here is to declare the extern "C" method without any
arguments, and we cast to the right signature when calling the function.