Add test case for `return` from rust!() macro
Addresses concerns raised in the discussion of changes in #93.

Generated code is complex and somehow different from original, so a
dedicated test case is needed to ensure that explicit return
statements (or rather expressions as what they technically are in Rust)
work in the same way as implicit ones, e.g. does not return too early
and doesn't break generated orchestrating code.

This test case is also implemented in a way that keeps clippy happy
about code style, so it does not raise new `needless_return` warnings.